# perf: greedy search for finding entry point