before:
```
level=7, ef_construct=30, ef=100 recall=0.9: construct=542.960s search=2087 us
```
after:
```
level=7, ef_construct=30, ef=100 recall=0.91: construct=546.145s search=1468 us
```